+++ docs/changes.rst
@@ -2,8 +2,8 @@ Changes
 =======
 
 
-Current
--------
+0.10.0 (23.03.2017)
+-------------------
 
 - added ``broadcast`` parameter to ``MicroService.emit_event``
 - added the ``broadcast`` parameter to ``BaseEventTransport.emit_event``